requirements.txt
streamlit>=1.28.0
openai>=0.28.0
google-generativeai>=0.3.0
requests>=2.31.0
pandas>=2.0.0
plotly>=5.18.0
matplotlib>=3.7.0
reportlab>=4.0.0
# Optional: Google Cloud Natural Language API
google-cloud-language>=2.11.0
# Optional: TextBlob (fallback if Vertex AI not available)
textblob>=0.17.1
# Optional: WordCloud
wordcloud>=1.9.2
python-dotenv>=1.0.0
